refactor(packages): change packages structure and rename some classes
