+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/config/ApplicationConfig.java
@@ -1,6 +1,6 @@
 package com.kopchak.worldoftoys.config;
 
-import com.kopchak.worldoftoys.user.UserRepository;
+import com.kopchak.worldoftoys.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
