+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/model/Role.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.user;
+package com.kopchak.worldoftoys.model;
 
 public enum Role {
     USER,
