+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/model/TokenType.java
@@ -1,0 +1,5 @@
+package com.kopchak.worldoftoys.model;
+
+public enum TokenType {
+    BEARER
+}
