+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/repository/TokenRepository.java
@@ -1,7 +1,9 @@
-package com.kopchak.worldoftoys.token;
+package com.kopchak.worldoftoys.repository;
 
 import java.util.List;
 import java.util.Optional;
+
+import com.kopchak.worldoftoys.model.Token;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
