+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/service/JwtTokenService.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.config;
+package com.kopchak.worldoftoys.service;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -15,7 +15,7 @@ import java.util.Map;
 import java.util.function.Function;
 
 @Service
-public class JwtService {
+public class JwtTokenService {
     private static final String SECRET_KEY = "404E635266556A586E3272357538782F413F4428472B4B6250645367566B5970";
 
     public String extractUsername(String token) {
