--- worldoftoys/src/main/java/com/kopchak/worldoftoys/token/TokenType.java
@@ -1,5 +1,0 @@
-package com.kopchak.worldoftoys.token;
-
-public enum TokenType {
-    BEARER
-}
